+++ react-carousel/src/plugins/clickToChange.js
