+++ react-carousel/src/plugins/slidesPerPage.js
